+++ data/examples/other/cpp/separation-1b-out.hs
@@ -4,7 +4,6 @@ decompressingPipe ::
   ConduitT ByteString ByteString m ()
 decompressingPipe Store = C.awaitForever C.yield
 decompressingPipe Deflate = Z.decompress $ Z.WindowBits (-15)
-
 #ifdef ENABLE_BZIP2
 decompressingPipe BZip2   = BZ.bunzip2
 #else