+++ data/examples/other/cpp/separation-2a-out.hs
@@ -1,5 +1,4 @@
 ffff, ffffffff :: Natural
-
 #ifdef HASKELL_ZIP_DEV_MODE
 ffff     = 200
 ffffffff = 5000